+++ src/game/main.js
@@ -1,6 +1,7 @@
 import { BootScene } from './scenes/BootScene'
 import { PreloadScene } from './scenes/PreloadScene'
 import { MenuScene } from './scenes/MenuScene'
+import { HistoryScene } from './scenes/HistoryScene'
 import { CharacterSelectScene } from './scenes/CharacterSelectScene'
 import { GameScene } from './scenes/GameScene'
 import { RewardScene } from './scenes/RewardScene'
@@ -22,6 +23,7 @@ const config = {
     BootScene,
     PreloadScene,
     MenuScene,
+    HistoryScene,
     CharacterSelectScene,
     GameScene,
     RewardScene,