+++ src/game/scenes/PreloadScene.js
@@ -44,6 +44,7 @@ export class PreloadScene extends Scene {
     this.load.image('bg-menu', 'backgrounds/fondoIntro.png')
     this.load.image('bg-characters', 'backgrounds/fondoPersonajes.png')
     this.load.image('bg-game', 'backgrounds/fondo_a.png')
+    this.load.image('bg-history', 'backgrounds/fondoHistory.png')
     this.load.image('boat', 'sprites/barco.png')
 
     // Sprites de personajes